refactor: clean up transaction history search and filter (#3284)
* refactor: Move Searchbar in own component and extract logic

* refactor: Move searchbar under header row + refactor search icon button

* fix: reset filter after clearing input

* chore: remove mock data

* Refactor: Create IconTextInput component for search bar

* chores: remove useless div
