+++ packages/shared/components/inputs/index.js
@@ -17,4 +17,6 @@ export { default as Radio } from './Radio.svelte'
 export { default as RecipientInput } from './RecipientInput.svelte'
 export { default as TextInput } from './TextInput.svelte'
 export { default as Toggle } from './Toggle.svelte'
+export { default as TogglableButton } from './TogglableButton.svelte'
+export { default as IconTextInput } from './IconTextInput.svelte'
 export { default as UnitInput } from './UnitInput.svelte'
