+++ packages/shared/components/panes/wallet/AccountActivityPane.svelte
@@ -1,0 +1,7 @@
+<script lang="typescript">
+    import { Pane, AccountActivity } from 'shared/components'
+</script>
+
+<Pane>
+    <AccountActivity />
+</Pane>
