+++ packages/shared/lib/core/wallet/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './formatBestMatchTokenAmount'
 export * from './formatDefaultTokenAmount'
 export * from './formatPreciseTokenAmount'
+export * from './searchTransactions'
